+++ src/util.ts
@@ -1,0 +1,22 @@
+import * as path from "path";
+// tslint:disable-next-line: no-implicit-dependencies
+import { Uri, workspace } from "vscode";
+import { IExtensionConfig } from "./types";
+
+export function getWorkspaceRelativePath(
+  filePath: string,
+  pathToResolve: string
+) {
+  if (workspace.workspaceFolders) {
+    const folder = workspace.getWorkspaceFolder(Uri.file(filePath));
+    return folder
+      ? path.isAbsolute(pathToResolve)
+        ? pathToResolve
+        : path.join(folder.uri.fsPath, pathToResolve)
+      : undefined;
+  }
+}
+
+export function getConfig(uri?: Uri): IExtensionConfig {
+  return workspace.getConfiguration("prettier", uri) as any;
+}